requirements.txt
# ============================================================
# Dépendances pour le projet Table-Detection
# Détecteur de tableaux dans des documents avec DETR
# ============================================================

# Core Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers et modèles pré-entraînés Hugging Face
transformers>=4.30.0
timm>=0.9.0

# Image Processing
Pillow>=9.0.0

# Hugging Face Hub (pour télécharger les modèles)
huggingface-hub>=0.16.0

# Utilities
numpy>=1.22.4,<2.3.0
scipy>=1.9.0

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0